refactor(template): 기존의 템플릿 탐색 API 와 템플릿 검색 API 를 구분하기 위해 기존의 템플릿 탐색 API 의 엔드포인트 변경
